# A few perf tweaks