Happened to notice a few things as I was looking at a trace of an app using this library...
- SimplePluralizer was using ToCharArray and then Any in both ToPluralInternal and ToSingularInternal. The ToCharArray is an expensive nop and can simply be removed. On .NET 8, Ascii.IsValid can be used instead of the Any call without loss of readability and is much faster.
- ODataNameMatchResolver is using Split().Last. It can just use LastIndexOf and Substring to avoid the string[] and strings for any other segments other than the last.
- HomogenizeEx can use a compiled / source generated regex.

Before

| Method    | Mean     | Allocated |
|---------- |---------:|----------:|
| Pluralize | 268.6 ns |      80 B |
| Resolve   | 213.1 ns |     320 B |

After

| Method    | Mean     | Allocated |
|---------- |---------:|----------:|
| Pluralize | 236.4 ns |      40 B |
| Resolve   | 124.1 ns |     256 B |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using Simple.OData.Client;

BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);

[MemoryDiagnoser(false)]
[HideColumns("Job", "Error", "StdDev", "Median", "RatioSD")]
public class Tests
{
    [Benchmark]
    public string Pluralize() => Pluralizers.Simple.Pluralize("runner");

    [Benchmark]
    public bool Resolve() => ODataNameMatchResolver.NotStrict.IsMatch("actualName", "requestedName");
}
```